refactor: flat util
