+++ weasel/util/__init__.py
@@ -1,12 +1,15 @@
-from .commands import download_file, get_checksum, get_git_version, get_hash
-from .commands import git_checkout, git_repo_branch_exists, join_command, run_command
-from .commands import split_command, upload_file, validate_project_commands
+from .commands import join_command, run_command, split_command
 from .config import load_project_config, parse_config_overrides
 from .config import substitute_project_variables
 from .environment import ENV_VARS, check_bool_env_var, check_spacy_env_vars
 from .filesystem import ensure_path, ensure_pathy, is_cwd, is_subpath_of, make_tempdir
 from .filesystem import working_dir
+from .frozen import SimpleFrozenDict, SimpleFrozenList
+from .git import _http_to_git, get_git_version, git_checkout, git_repo_branch_exists
+from .git import git_sparse_checkout
+from .hashing import get_checksum, get_hash
 from .logging import logger
 from .modules import import_file
-from .structures import SimpleFrozenDict, SimpleFrozenList
+from .remote import download_file, upload_file
+from .validation import validate_project_commands
 from .versions import get_minor_version, is_compatible_version, is_minor_version_match
