+++ weasel/util/frozen.py
