+++ weasel/util/git.py
@@ -4,8 +4,8 @@ from typing import Tuple
 
 from wasabi import msg
 
-from ...util.filesystem import is_subpath_of, make_tempdir
-from .run import run_command
+from .commands import run_command
+from .filesystem import is_subpath_of, make_tempdir
 
 
 def git_checkout(
