+++ weasel/util/hashing.py
