+++ weasel/util/remote.py
